requirements.txt
# Core dependencies
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0

# LLM clients
groq>=0.4.0
# openai>=1.0.0  # Uncomment when adding OpenAI support
# requests>=2.31.0  # For vLLM client

# Optional: deduplication
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Optional: testing
pytest>=7.4.0

#UI-demo
streamlit